+++ CHANGES.md
@@ -1,4 +1,8 @@
+Master
+------------------
+
 * atdgen: use Yojson 2.0 API (#299)
+* atdts: fix nullable object field writer (#312)
 
 2.9.1 (2022-06-10)
 ------------------