+++ CHANGES.md
@@ -1,6 +1,11 @@
 Master
 ------------------
 
+* atdpy: Support recursive definitions
+
+Master
+------------------
+
 * atdts: fix nullable object field writer (#312)